send email notifications on task failures
