+++ compose.yaml
@@ -102,6 +102,7 @@ x-airflow-common:
     AIRFLOW_VAR_HONEYBADGER_API_KEY: ${HONEYBADGER_API_KEY}
     AIRFLOW_VAR_HONEYBADGER_ENV: ${HONEYBADGER_ENV}
     AIRFLOW_VAR_CLEANUP_INTERVAL_DAYS: 10
+    AIRFLOW_VAR_EMAIL_ADDRESS_FOR_ERRORS: rialto-team@lists.stanford.edu
   volumes:
     - ${AIRFLOW_PROJ_DIR:-.}/rialto_airflow:/opt/airflow/rialto_airflow
     - ${AIRFLOW_PROJ_DIR:-.}/logs:/opt/airflow/logs
